# Make AttributeValue accessors to return error in case of type mismatch