… model/internal/pdata/common_test.go