This is an alternative solution to https://github.com/open-telemetry/opentelemetry-collector/pull/4909 proposed by @Aneurysm9

The following methods are deprecated in favor of the new accessors:
- StringVal -> StringValue
- IntVal -> IntValue
- DoubleVal -> DoubleValue
- BoolVal -> BoolValue
- MapVal -> MapValue
- SliceVal -> SliceValue
- BytesVal -> BytesValue

The new accessors return additional error. The error is raised if accessor is called for a wrong type, e.g. `IntValue()` is called on `AttributeValue` of `AttributeValueTypeString` type.

Corresponding setter methods are also renamed for consistency without changing its logic:
- SetStringVal -> SetStringValue
- SetIntVal -> SetIntValue
- SetDoubleVal -> SetDoubleValue
- SetBoolVal -> SetBoolValue
- SetBytesVal -> SetBytesValue

## Benchmarks

goos: darwin
goarch: amd64
pkg: go.opentelemetry.io/collector/model/internal/pdata
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz

### Before

```
BenchmarkAttributeValueStringDeprecatedAccessor-16                  1000000000           0.6064 ns/op          0 B/op          0 allocs/op
BenchmarkAttributeValueIntDeprecatedAccessor-16                     1000000000           0.6103 ns/op          0 B/op          0 allocs/op
BenchmarkAttributeValueDoubleDeprecatedAccessor-16                  1000000000           0.6048 ns/op          0 B/op          0 allocs/op
BenchmarkAttributeValueBoolDeprecatedAccessor-16                    1000000000           0.6933 ns/op          0 B/op          0 allocs/op
BenchmarkAttributeValueBytesDeprecatedAccessor-16                   1000000000           0.6501 ns/op          0 B/op          0 allocs/op
BenchmarkAttributeValueMapDeprecatedAccessor-16                     556089897            2.182 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueArrayDeprecatedAccessor-16                   541185199            2.179 ns/op           0 B/op          0 allocs/op

```

### After

```
BenchmarkAttributeValueStringAccessor-16                            353806957            3.464 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueIntAccessor-16                               346630108            3.479 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueDoubleAccessor-16                            356147574            3.358 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueBoolAccessor-16                              350702415            3.472 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueBytesAccessor-16                             331781389            3.677 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueMapAccessor-16                               341309820            3.559 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueArrayAccessor-16                             346380688            3.493 ns/op           0 B/op          0 allocs/op
```