@dmitryax 2022-03-13
I see some pros and cons of this approach comparing to https://github.com/open-telemetry/opentelemetry-collector/pull/4909

Pros:

- Safe usage, no chance to catch a runtime error.

Cons:

- Clunky interface raising only one error which can be omitted by user if the type is checked in advance. We have to explicitly communicate this to assure client that there is no other error can be returned.
- Degraded performance.

If we decide to go with this approach we have to update : 
- [numeric value field accessors](https://github.com/open-telemetry/opentelemetry-collector/pull/4980) and 
- [Metric data field accessors](https://github.com/open-telemetry/opentelemetry-collector/blob/main/model/internal/pdata/generated_metrics.go#L630-L663)

Both of them will likely introduce the performance hit as well.