@bogdandrutu 2022-03-13
I don't like this especially because of the extra check when I already know the type. we should go with some standard way of doing this. 

1. For example let's have the same behavior as Google protobuf https://github.com/open-telemetry/opentelemetry-proto-go/blob/ee3f9617238186538f35760934fc1727e807b5c3/otlp/common/v1/common.pb.go#L131 unfortunately no easy way to check for nil on the struct case.
2. Panic, this very similar to a type cast, and if you don't check the conversion status you get panic. So not sure why doing this. I read the comments in the other PR but don't prove to me why panic is not natural especially how type casting works in golang