@dmitryax 2022-03-17
Closing based on decision made in https://github.com/open-telemetry/opentelemetry-collector/issues/4775